Fix nested Aggregations. Make NcmlElementReader a subclass instead of an inner class of NcmlReader. Cleanup the trueUrl of nested Aggregations. MFiles.create doesnt allow null location.
