+++ cdm-core/src/main/java/ucar/nc2/internal/ncml/package-info.java
@@ -1,0 +1,7 @@
+/*
+ *  Copyright (c) 1998-2021 John Caron and University Corporation for Atmospheric Research/Unidata
+ *  See LICENSE for license information.
+ */
+
+/** Implementation of NcML. */
+package ucar.nc2.internal.ncml;
